telegram_bot/requirements.txt
python-telegram-bot==20.8
pymongo==4.6.1
pysftp==0.2.9
paramiko==3.4.0
beautifulsoup4==4.12.3
html5lib==1.1
requests==2.31.0
python-dotenv==1.0.1
validators==0.22.0
pillow==10.2.0
cryptography==42.0.7 